Ajout des couts d'amelioration et ajout des stats d'amelioration

+++ VirusDefense/src/main/java/fr/iut/virusdefense/modele/cellules/attaque/AtkRayon.java
@@ -8,7 +8,7 @@ import fr.iut.virusdefense.modele.maladies.Maladie;
 
 public abstract class AtkRayon extends Attaque {
 
-    public AtkRayon(Cellule cellule, int degats){
+    public AtkRayon(Cellule cellule, double degats){
         super(cellule, degats);
     }
