+++ VirusDefense/src/main/java/fr/iut/virusdefense/modele/entitesgeneriques/Zone.java
@@ -9,13 +9,13 @@ import java.util.List;
 
 public class Zone extends EntiteAtk{
     private ArrayList<Maladie> cibles;
-    private final int degats;
+    private final double degats;
     private int age;
     private final int ageMaximal;
     private final double rayonZone;
 
 
-    public Zone(Entite e1, ArrayList<Maladie> cibles, int degats, int ageMaximal, List<Alteration> alterations, double portee) {
+    public Zone(Entite e1, ArrayList<Maladie> cibles, double degats, int ageMaximal, List<Alteration> alterations, double portee) {
         super(e1.getEnvironnement(), e1.getLigne(), e1.getColonne(), alterations);
         age = 0;
         this.ageMaximal = ageMaximal;
