+++ VirusDefense/src/main/java/fr/iut/virusdefense/modele/maladies/Maladie.java
@@ -11,7 +11,7 @@ import java.util.List;
  * Représente une maladie
  */
 public abstract class Maladie extends Entite {
-    private int pv;
+    private double pv;
     private final double vitesse;
     private final int recompense;
     private double coefVitesse;
@@ -46,7 +46,7 @@ public abstract class Maladie extends Entite {
         this.pv = 0;
     }
 
-    public int getPv() {
+    public double getPv() {
         return pv;
     }
 
@@ -54,7 +54,7 @@ public abstract class Maladie extends Entite {
      * Méthode qui calcule les PV de la maladie apres avoir subis des degats de la cellule
      * @param degats degats subis par la maladie
      */
-    public void prendreDegats(int degats){
+    public void prendreDegats(double degats){
         if (degats > 0)
             pv -= degats;
     }
@@ -93,7 +93,7 @@ public abstract class Maladie extends Entite {
     }
 
     public void infligerDegatsAuJoueur(){
-        getEnvironnement().getJoueur().retirerPv(pv);
+        getEnvironnement().getJoueur().retirerPv((int) pv);
     }
